safari-600.7-branch/LayoutTests/java/lc3/JSNumber/ToChar-003-n-expected.txt

LiveConnect 3.0 JavaScript to Java Data Type Conversion number conversion to char

PASS dt.setChar(java.lang.Character.MIN_VALUE - 1); dt.PUB_CHAR threw exception ReferenceError: Can't find variable: java.
PASS dt.setChar(java.lang.Character.MIN_VALUE - 1); dt.getShort() threw exception ReferenceError: Can't find variable: java.
PASS typeof dt.getShort() is 'number'
PASS successfullyParsed is true

TEST COMPLETE
 
